Fix field width specifier in gio/giomodule.c
gio/giomodule.c: In function ‘print_help’:
gio/giomodule.c:727:23: error: field width specifier ‘*’ expects argument of type ‘int’, but argument 2 has type ‘gsize’ {aka ‘long unsigned int’} [-Werror=format=]
           g_print (" %*s - %d\n", width, g_io_extension_get_name (extension), g_io_extension_get_priority (extension));
                      ~^~          ~~~~~

+++ gio/giomodule.c
@@ -724,7 +724,7 @@ print_help (const char        *envvar,
         {
           extension = l->data;
 
-          g_print (" %*s - %d\n", width, g_io_extension_get_name (extension), g_io_extension_get_priority (extension));
+          g_print (" %*s - %d\n", (int) width, g_io_extension_get_name (extension), g_io_extension_get_priority (extension));
         }
     }
 }
